+++ catsoop/__AUTH__/login/login.py
@@ -293,22 +293,6 @@ def get_logged_in_user(context):
         entered_password = form.get("cs_hashed_0", "")
 
         valid_uname = True
-        if _validate_email(context, uname) is None:
-            # this looks like an e-mail address, not a username.
-            # find the associated username, if any
-            # TODO: implement caching of some kind so this isn't so slow/involved
-            data_root = context.get("cs_data_root", base_context.cs_data_root)
-            global_log_dir = os.path.join(data_root, "_logs")
-            for d in os.listdir(global_log_dir):
-                if not d.endswith(".db"):
-                    continue
-                u = d[:-3]
-                e = logging.most_recent("_logininfo", [], u, {})
-                e = e.get("email", None)
-                if e == uname:
-                    uname = u
-                    break
-
         vmsg = _validate_username(context, uname)
         if vmsg is not None:
             valid_uname = False