+++ CHANGELOG.md
@@ -26,6 +26,9 @@ _Work toward next release. Currently under development._
 
 **REMOVED:**
 
+* Removed slow/broken code for logging in via e-mail instead of username when
+    using the `login` auth type.
+
 **FIXED:**
 
 * Show an error message when two questions would have the same name (#53)